Fix compile warning on win.
--HG--
branch : trunk

+++ src/shared/Threading.h
@@ -40,7 +40,7 @@ namespace ACE_Based
                     delete this;
             }
         private:
-            ACE_Atomic_Op<ACE_Thread_Mutex, int> m_refs;
+            ACE_Atomic_Op<ACE_Thread_Mutex, long> m_refs;
     };
 
     enum Priority
